+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Animals movement speed when attacking is now configurable
+
 ## Beta 1.3.0
 * Added Animal Module
   * Added Animal Attacking Feature. Make animals fight back and no longer flee when attacked